refactor: 게시글 삭제 기능 findById -> existsById 변경

+++ backend/src/main/java/cotato/backend/domains/post/PostService.java
@@ -85,8 +85,9 @@ public class PostService {
 
 	@Transactional
 	public void deletePostById(Long postId) {
-		Post post = postRepository.findById(postId)
-			.orElseThrow(() -> ApiException.from(POST_NOT_FOUND));
+		if (!postRepository.existsById(postId)) {
+			throw ApiException.from(POST_NOT_FOUND);
+		}
 		postRepository.deleteById(postId);
 	}
 }
